Set sensible permissions (0644) for the output file
By default, exelize saves the file with the executable bit set. This is
unnecessary for a spreadsheet, and a more sensible set of permissions is
0644. Unfortunately the saveAs function does not have an option for
saving the file with a specific set of permissions, so we have to save
the file first and then change the permissions using os.Chmod.

Thanks to @jpmens for reporting this

Closes #5