Switch to Cobra for command line argument parsing
Cobra supports more functionality than the standard library flags
module, including:

Mandatory arguments
Sub commands

This will allow more flexibility for future development, as well as
simplifying the code.

The password flag is now optional and the user will only be prompted for
the password if the flag is set.

The hostname flag is now optional and defaults to localhost, in line
with the MySQL / MariaDB command line tools.

The other flags have been updated so that their short and long forms
match the MySQL / MariaDB command line tools. However, there is no short
option for the hostname (-h) as this conflicts with the help command.

Closes #9
Closes #6
Closes #4