Allow table names to be specified when exporting
The export subcommand now supports a --table option, which can be
passed zero or more times. Zero times means 'include all tables', and
this is the default behaviour. If one or more tables are specified, only
those tables will be included in the export.

There is currently no error displayed if any of the table names do not
exist.

Closes #3